% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_stats.R
\name{closest_match}
\alias{closest_match}
\title{Match list of standards to peak table}
\usage{
closest_match(stds, peakTable, rt_tol = 0.25, mz_ppm = 30, rt_col = "rt",
  mz_col = "mz", int_col = "into")
}
\arguments{
\item{stds}{\code{\link{tibble}} of standards to match to a peak table}

\item{peakTable}{\code{\link{tibble}} containing peak table supplied by \code{\link{findPeaks}} (but converted to \code{\link{tibble}}/\code{\link{data.frame}}).}

\item{rt_tol}{Retention time tolerance for matching peaks. Pay attention to the unit of your tables.
rt_tol should match and stds and peakTable should use same units (i.e. minutes of seconds).}

\item{mz_ppm}{ppm for matching peaks.}

\item{rt_col}{Character string giving the column containing the retention times. Must be same in standards and peak table.}

\item{mz_col}{Character string giving the column containing the m/z values. Must be same in standards and peak table.}

\item{int_col}{Character string giving the column containing the intensities in the peak table.}
}
\value{
A vector having the length equivalent to the number of rows in stds giving the indices of the hits in peakTable.
}
\description{
This function will match a table of standard compounds and a peak table by m/z and retention time.
If there is more than one possible hit the highest intensity peak will be chosen.
}
