% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_operations.R
\name{dbPool_MetabolomiQCs}
\alias{dbPool_MetabolomiQCs}
\title{Write message to QC4Metabolomics db}
\usage{
dbPool_MetabolomiQCs(idleTimeout = 1)
}
\arguments{
\item{idleTimeout}{The number of minutes that an idle object will be kept in the pool before it is destroyed.}
}
\value{
A \code{\link{dbPool}} object to connect to the MetabolomiQC database using settings in the ini file.
}
\description{
Write message to QC4Metabolomics db
}
\keyword{internal}
