% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_extract.R
\name{EIC_calc}
\alias{EIC_calc}
\title{Calculate EIC from a raw matrix}
\usage{
EIC_calc(tbl, lower, upper, BPI = FALSE)
}
\arguments{
\item{tbl}{A \code{\link{tibble}} containing the columns: 
\itemize{
  \item \strong{scan:} scan number
  \item \strong{scan_rt:} Retention time of scan
  \item \strong{intensity:} The intensity of the observation
  \item \strong{mz:} the mz of the observation
}}

\item{lower}{Lower boundary of EIC slice}

\item{upper}{Upper boundary of EIC slice}

\item{BPI}{Logical selecting to calculate TIC (FALSE) or BPI.}
}
\value{
tbl A \code{\link{tibble}} containing the columns: 
\itemize{
  \item \strong{scan:} scan number
  \item \strong{scan_rt:} Retention time of scan
  \item \strong{intensity:} The summed intensity for each scan in the given m/z interval
}
}
\description{
Calculate EIC from a raw matrix of all observations (scan/mz/intensity combinations).
}
\keyword{internal}

