% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contaminants_untargeted.R
\name{EIC_contaminants}
\alias{EIC_contaminants}
\title{Find EICs that behave like contaminants}
\usage{
EIC_contaminants(raw, bin_ppm = 30, interval_ppm = 30, min_time = 5,
  merge_corr = 0.9, merge_ppm = 30, min_int = 5000)
}
\arguments{
\item{raw}{\code{\link{xcmsRaw}} object to profile}

\item{bin_ppm}{Tolerance (ppm) for initial binning of m/z values}

\item{interval_ppm}{Tolerance for creating final EICs after merging similar bins}

\item{min_time}{Minimum time (minutes) an EIC should be above \code{min_int} to be considered a contaminant.}

\item{merge_corr}{Minimum correlation between EICs to be merged.}

\item{merge_ppm}{Maximum difference (ppm) between EICs to be merged.}

\item{min_int}{Minimum intensity that the EIC needs to be above for a a minimum of \code{min_time}.}
}
\value{
A \code{\link{tibble}} containing the columns: 
\itemize{
  \item \strong{mz:} m/z of the proposed contaminant
  \item \strong{EIC:} EIC of the m/z.
}
}
\description{
This function looks in raw LC-MS data for "features"/EICs that behave like contaminants.
Behaving like contaminants in this case means that a certain m/z values is 
present in more than \code{min_time} above intensity \code{min_int}.
}

