% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_cont_list}
\alias{get_cont_list}
\title{Get list of known contaminants}
\usage{
get_cont_list(polarity = c("positive", "negative", "unknown"), type = "URL")
}
\arguments{
\item{polarity}{The polarity to get contaminants for. Can be "positive", "negative" or "unknown".
If "unknown" the list specified in the MetabolomiQCs.conf is used. 
MetabolomiQCs.conf can be in the working folder or the home folder.
If those are not found the package default is used (unknown will used the positive mode list).}

\item{type}{If using local or remote. 
Only "URL" implemented which downloads a list from https://github.com/stanstrup/common_mz}
}
\value{
tbl A \code{\link{tibble}} containing the columns: 
\itemize{
  \item \strong{Monoisotopic ion mass (singly charged):} m/z of the contaminant
  \item \strong{Ion type:} Notation for adduct/fragment type
  \item \strong{Formula for M or subunit or sequence:} Molecular formula
  \item \strong{Compound ID or species:} Name of the compound
  \item \strong{Possible origin and other comments:} Suggestion for the origin of the contaminant
  \item \strong{References:} Reference for the contaminant
}
}
\description{
Get list of known contaminants
}
\references{
https://github.com/stanstrup/common_mz
}
