% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_extract.R
\name{get_EICs}
\alias{get_EICs}
\title{Get EICs from xcmsRaw object}
\usage{
get_EICs(xraw, range_tbl, exclude_mz = NULL, exclude_ppm = 30,
  range_tbl_cols = c("mz_lower", "mz_upper"), BPI = FALSE)
}
\arguments{
\item{xraw}{\code{\link{xcmsRaw}} object to get EIC(s)/TIC from.}

\item{range_tbl}{data.frame/\code{\link{tibble}} with columns for the lower and upper m/z boundaries of EIC slice(s).}

\item{exclude_mz}{Masses to exclude from the EIC. Most useful to remove contaminants from TICs.}

\item{exclude_ppm}{ppm tolerance of exclude_mz}

\item{range_tbl_cols}{Which columns in range_tbl holds the lower and upper range. defaults to c("mz_lower","mz_upper").}

\item{BPI}{Logical selecting to calculate TIC (FALSE) or BPI.}
}
\value{
tbl A \code{\link{tibble}} containing the columns: 
\itemize{
  \item \strong{scan:} scan number
  \item \strong{scan_rt:} Retention time of scan
  \item \strong{intensity:} The summed intensity for each scan in the given m/z interval
}
}
\description{
Takes an \code{\link{xcmsRaw}} object and extracts EICs.
Can do multiple ranges and exclude certain masses unlike \code{\link{getEIC}}.
Can be used to extract the TIC too.
}

