% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_QC4Metabolomics_settings}
\alias{get_QC4Metabolomics_settings}
\title{Get table with QC4metabolomics settings}
\usage{
get_QC4Metabolomics_settings(modules = NULL)
}
\arguments{
\item{modules}{character strong. Get settings only for selected modules.}
}
\value{
tbl A \code{\link{tibble}} containing the columns: 
\itemize{
  \item \strong{name:} name of the setting
  \item \strong{value:} value of the settin
  \item \strong{module:} which module the setting belongs to.
}
}
\description{
Get table with QC4metabolomics settings
}
