% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EIC.R
\name{getEIC_C_wrap}
\alias{getEIC_C_wrap}
\title{Calculate EICs from a raw matrix using XCMS C function}
\usage{
getEIC_C_wrap(xraw_values, range_tbl)
}
\arguments{
\item{range_tbl}{data.frame/\code{\link{tibble}} with columns for the lower and upper ("mz_lower","mz_upper") m/z boundaries of EIC slice(s).}

\item{tbl}{A \code{\link{tibble}} containing the columns: 
\itemize{
  \item \strong{scan:} scan number
  \item \strong{scan_rt:} Retention time of scan
  \item \strong{intensity:} The intensity of the observation
  \item \strong{mz:} the mz of the observation
}}
}
\value{
tbl A \code{\link{tibble}} containing the columns: 
\itemize{
  \item \strong{scan:} scan number
  \item \strong{scan_rt:} Retention time of scan
  \item \strong{intensity:} The summed intensity for each scan in the given m/z interval
}
}
\description{
Calculate EICs from a raw matrix of all observations (scan/mz/intensity combinations).
Can calculate for several ranges at a time
}
\keyword{internal}
