% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icmeters.R
\name{ic_boxes}
\alias{ic_boxes}
\title{Convert raw data into a tibble of xcmsRaw objects.}
\usage{
ic_boxes(token)
}
\arguments{
\item{token}{A character vector containing the ICMeter token}
}
\value{
data.frame containing information about each box registered on the token account
}
\description{
Convert raw data into a tibble of xcmsRaw objects.
}
