% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icmeters.R
\name{ic_token}
\alias{ic_token}
\title{Convert raw data into a tibble of xcmsRaw objects.}
\usage{
ic_token(user, pass)
}
\arguments{
\item{user}{ICMeter username}

\item{pass}{ICMeter password}
}
\value{
A character vector containing the ICMeter token
}
\description{
Convert raw data into a tibble of xcmsRaw objects.
}
