% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_filenames.R
\name{parse_filenames}
\alias{parse_filenames}
\title{Parse filenames from given string}
\usage{
parse_filenames(filenames, filenamemask)
}
\arguments{
\item{filenames}{Character string with the filenames to parse.}

\item{filenamemask}{Filename mask. Variables to capture are inclosed by "\%". See examples}
}
\value{
data.frame with fields parsed from the filenames
}
\description{
Parse filenames from given string
}
\examples{
testfilenames <- c("RemoveThis_METSY_sample1___17-random_AndThis.txt",
                   "RemoveThis_METSY_sample2___21-something_AndThis.txt",
                   "RemoveThis_METSY_sample3___123-blah_AndThis.txt",
                   "RemoveThis_JDRF_SameName___1-dummyA_AndThis.txt",
                   "RemoveThis_JDRF_SameName___2-dummyB_AndThis.txt",
                   "RemoveThis_JDRF_SameName___3-dummyC_AndThis.txt",
                   "RemoveThis_this_file___name-is_wrong",
                   "RemoveThis_and_this_AndThis.txt",
                   "RemoveThis_this_one___should-work",
                   "RemoveThis_how_about___this_-_one_AndThis.txt"
                  )

defaultmask   <- "RemoveThis_\%study\%_\%name\%___\%rep\%-\%dummy\%_AndThis.txt"

parse_filenames(testfilenames,defaultmask)


}
\author{
Stinus Lindgreen, \email{sldg@steno.dk}; Jan Stanstrup, \email{jpzs@steno.dk}
}

