% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_stats.R
\name{peak_factor}
\alias{peak_factor}
\title{Calculate Tailing Factor and Asymmetry Factor}
\usage{
peak_factor(EIC, rt, factor = "TF")
}
\arguments{
\item{EIC}{EIC containing the peak to calculate for. 
\code{\link{tibble}} as produced with \code{\link{get_EICs}}.}

\item{rt}{Retention time of the center of the peak (Numeric)}

\item{factor}{to calculate. Character string either "TF" (Tailing Factor) or "ASF" (Asymmetry Factor).}
}
\value{
Numeric
}
\description{
Calculate Tailing Factor and Asymmetry Factor
}
\references{
http://www.chromforum.org/viewtopic.php?t=20079
}

