% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_chrom}
\alias{plot_chrom}
\title{Plot chromatogram}
\usage{
plot_chrom(tbl, RT_col = "RT", Intensity_col = "Intensity")
}
\arguments{
\item{tbl}{tbl with retention time and intensity to plot}

\item{RT_col}{Name of the rentention time column}

\item{Intensity_col}{Name of the intensity column}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Plot chromatogram
}
