% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_contaminants}
\alias{plot_contaminants}
\title{Bar plot of contaminants}
\usage{
plot_contaminants(data, title, x_var = "comp_name", y_var = "EIC_median")
}
\arguments{
\item{data}{tbl with the contamination amounts}

\item{title}{Plot title}

\item{x_var}{Column name that holds the compound/contaminant names}

\item{y_var}{Column name that holds the compound/contaminant values}
}
\value{
a \code{\link{ggplot}} object.
}
\description{
Bar plot of contaminants
}
