% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plotly_clean_tt}
\alias{plotly_clean_tt}
\title{Make replacements in plotly tooltips}
\usage{
plotly_clean_tt(plotly, rep)
}
\arguments{
\item{plotly}{A \code{\link[plotly]{plotly}} object.}

\item{rep}{A named character vector. Names are the text to replace and the string is the replacement string.}
}
\value{
A \code{\link[plotly]{plotly}} object.
}
\description{
Make replacements in plotly tooltips
}

