% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_operations.R
\name{rem_dead_files}
\alias{rem_dead_files}
\title{Remove files that no longer exist from the database}
\usage{
rem_dead_files(file_md5, path, pool = NULL, log_source)
}
\arguments{
\item{file_md5}{A vector giving the md5 of the files to check.}

\item{path}{A vector giving the relative path to the files to check.}

\item{pool}{(see pool package) to use for writing. If null a new connection will be made reading the connection details from the conf file.}
}
\value{
Database change and a logical vector saying which files existed.
}
\description{
Remove files that no longer exist from the database
}
\keyword{internal}
