% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{set_QC4Metabolomics_settings_from_file}
\alias{set_QC4Metabolomics_settings_from_file}
\title{Set QC4metabolomics settings form env file}
\usage{
set_QC4Metabolomics_settings_from_file(file)
}
\arguments{
\item{file}{character string with the filename of the env file}
}
\description{
Set QC4metabolomics settings form env file
}
