% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_stats.R
\name{tbl2ROI}
\alias{tbl2ROI}
\title{Convert a list of peaks (rt / m/z pairs) to a Region of Interest (ROI) list for use with \code{\link{findPeaks}}}
\usage{
tbl2ROI(tbl, raw, ppm, rt_tol)
}
\arguments{
\item{tbl}{\code{\link{tibble}} containing the columns "rt" and "mz". rt needs to be in seconds.}

\item{raw}{\code{\link{xcmsRaw}} object to create ROI for. It needs to be a specific \code{\link{xcmsRaw}} to match retention times to scan nubmers.}

\item{ppm}{ppm tolerance for the generated ROI.}

\item{rt_tol}{Retention time tolerance (in sec!) for the generated ROI.}
}
\value{
List containing the ROIs. Each list contains mz, mzmin, mzmax, scmin, scmax, length (set to -1, not used by centWave) and intensity (set to -1, not used by centWave) columns.
}
\description{
Convert a list of peaks (rt / m/z pairs) to a Region of Interest (ROI) list for use with \code{\link{findPeaks}}
}
