% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_operations.R
\name{write_to_log}
\alias{write_to_log}
\title{Write message to QC4Metabolomics db}
\usage{
write_to_log(msg, cat, source, pool = NULL)
}
\arguments{
\item{msg}{The message. A character vector.}

\item{cat}{The category ("info", "warning" or "error"). A character vector.}

\item{pool}{(see pool package) to use for writing. If null a new connection will be made reading the connection details from the conf file.}
}
\value{
Nothing. Written to db.
}
\description{
Write message to QC4Metabolomics db
}
\keyword{internal}

