% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_to_tbl.R
\name{xcmsRaw_to_tbl}
\alias{xcmsRaw_to_tbl}
\title{Convert raw data into a tibble of xcmsRaw objects.}
\usage{
xcmsRaw_to_tbl(files, ...)
}
\arguments{
\item{files}{character vector of file names/paths.}
}
\value{
A \code{\link{tibble}} containing the columns: 
\itemize{
  \item \strong{file:} Filename without path.
  \item \strong{polarity:} Character string of "positive", "negative", or "unknown".
  \item \strong{raw:} The xcmsRaw objects.
  \item \strong{path:} The input path (files).
}
}
\description{
Convert raw data into a tibble of xcmsRaw objects.
}

